mysql: always import driver
When under GOOS=wasip1, the github.com/go-sql-driver/mysql driver
is imported and registered with the database/sql package. When
not under GOOS=wasip1, the package isn't imported and so isn't
registered.

This commit updates the mysql package to always import the driver
and always register it with database/sql.